Ensure ocamlformat-rpc is available in dev flake
A recent version of ocamlformat is included in the nativeBuildInputs of
the mina#with-lsp dev flake to ensure that ocamlformat-rpc is available
for editors in the flake environment. The ocamlformat package that was
formerly included in the flake was 0.20.1, and that version did not
bundle ocamlformat-rpc with it. Only versions 0.22.4 and later of
ocamlformat include ocamlformat-rpc.